# Improve array decoding performance