Hi @gfx I've used this msgpack library in my project and found small optimization that improves decoding speed by 3-4%.