@sergeyzenchenko 2019-05-27
For large arrays performance difference can be even bigger and also it will produce less trash for GC