@gfx 2019-05-28
Tested with this benchmark script:

```typescript
/* eslint-disable no-console */
import { encode, decode } from "../src";
// @ts-ignore
import _ from "lodash";

const data = _.cloneDeep(new Array(1000).fill([[1], [1, 2, 3], [1, 2, 3, 4, 5]]));

// warm up
const encoded = encode(data);
decode(encoded);

// run

console.time("encode");
for (let i = 0; i < 10000; i++) {
  encode(data);
}
console.timeEnd("encode");

console.time("decode");
for (let i = 0; i < 10000; i++) {
  decode(encoded);
}
console.timeEnd("decode");
```

And this PR significantly improves decoding:

before:

```
$ npx ts-node benchmark/sandbox.ts
encode: 2753.986ms
decode: 1855.263ms
```

after:

```
$ npx ts-node benchmark/sandbox.ts
encode: 2393.308ms
decode: 3804.347ms
```

Great!