@gfx 2019-05-28
(CI failed because of SauceLabs credentials; will fix it later https://docs.travis-ci.com/user/environment-variables/ )