@nelsonlaquet 2019-07-02
I wonder if you could improve the performance further on certain engines by allocating an array without holes in it:

```javascript
array: Array.from({ length: size }),
```

Source:
https://v8.dev/blog/elements-kinds
https://2ality.com/2018/12/creating-arrays.html