@gfx 2019-07-03
Tried it now, but `Array.from({ length: size })` makes decoding 10x slower than `new Array(size)` :-(